refactor: rename Adapter/index.ts and Balances/index.ts files to useAdapter.ts and useBalances.ts
